move test to appimagetest.ts
